make QRbooking package
