+++ QRBooking/QREventBookings.egg-info/SOURCES.txt
@@ -5,3 +5,4 @@ QRBooking/QREventBookings.egg-info/SOURCES.txt
 QRBooking/QREventBookings.egg-info/dependency_links.txt
 QRBooking/QREventBookings.egg-info/top_level.txt
 QRBooking/generateOR/__init__.py
+QRBooking/generateOR/pdf.py
